feat: new fields sensor

+++ src/app/controllers/AlertController/remove.js
@@ -1,8 +1,17 @@
 import Alert from '../../models/Alert';
+import { deleteSubscription } from '../../services/subscriptions';
 
 export default async (req, res) => {
   try {
-    // TO DO (DELETE SUBSCRIPTION SERVICE - fiwareSubscriptionId)
+    const alert = await Alert.findOne({
+      _id: await req.params.id.toString(),
+    });
+
+    if (!alert) {
+      return res.status(400).json({ error: 'Alert not found!' });
+    }
+
+    await deleteSubscription(alert.fiwareSubscriptionId);
 
     const alertRmv = await Alert.deleteOne({
       _id: await req.params.id.toString(),
